linux bullshit type fixing

+++ Core/SoarKernel/src/SQLiteCpp/Column.cpp
@@ -65,15 +65,15 @@ unsigned Column::getUInt() const noexcept // nothrow
 }
 
 // Return the 64bits integer value of the column specified by its index starting at 0
-long long Column::getInt64() const noexcept // nothrow
+int64_t Column::getInt64() const noexcept // nothrow
 {
     return sqlite3_column_int64(mStmtPtr, mIndex);
 }
 
 // Return the 64bits unsigned integer value of the column specified by its index starting at 0
-unsigned long long Column::getUInt64() const noexcept // nothrow
+uint64_t Column::getUInt64() const noexcept // nothrow
 {
-    return static_cast<unsigned>(sqlite3_column_int64(mStmtPtr, mIndex));
+    return static_cast<uint64_t>(sqlite3_column_int64(mStmtPtr, mIndex));
 }
 
 // Return the double value of the column specified by its index starting at 0
